fix integer overflow/underflow in exphist corner case

+++ sdk/metrics/src/main/java/io/opentelemetry/sdk/metrics/internal/state/AdaptingCircularBufferCounter.java
@@ -73,13 +73,13 @@ public class AdaptingCircularBufferCounter implements ExponentialCounter {
 
     if (index > endIndex) {
       // Move end, check max size
-      if (index - startIndex + 1 > backing.length()) {
+      if ((long) index - (long) startIndex + 1L > backing.length()) {
         return false;
       }
       endIndex = index;
     } else if (index < startIndex) {
       // Move end, check max size
-      if (endIndex - index + 1 > backing.length()) {
+      if ((long) endIndex - (long) index + 1L > backing.length()) {
         return false;
       }
       startIndex = index;
